+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `-f/--follow` option to `gantry logs`.
+- Added `--python-version` option to `gantry run` for overriding the default Python version to use.
 
 ### Changed